requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlmodel
pydantic-settings
python-dotenv
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
psycopg2-binary
email-validator
pytest
pytest-asyncio
httpx
jinja2
tzdata

